requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
PyJWT==2.8.0
gunicorn==21.2.0
flask-limiter
psycopg2-binary==2.9.9
python-dotenv==1.0.0 